Added Login Page
The page has been added alongside connecting its static files.
The page has also been connected with the landing page and the
register page. Additionally in the register pages form, I set up
the action url and method. Next thing to do is to fix a bug in the login
pages JS. The toggle show password feature does not work.