Bug Fix!
The issue with the toggle show password button was simply due
to the js script not being properly connected to the login.html.
The scripts path was {& static 'login.js' %} when it was instead supposed
to be {% static 'base/login.js' %}. The bug has been fixed.
I did however fix an issue in the css with the toggle show password not
displaying the proper icon on startup. It has been fixed.