Bug Fix!
In the register an account page. The form was not submitting at all.
This was due to an invalid line in the js and the browser, not being
cached properly. It has been fixed. And I additionally removed that
line from the login.js as well to prevent that issue from arising again.