Added Account Created Card To Login
Now when the user successfully creates an account they are redirected
to the login page which also receives a created=true query parameter,
the login page checks for this and if it finds it, it displays the
account created card. The card has a close button if users want to close
it.