Changed JS validation
I have changed the max char length for meetings in the JS
configurations. Now titles have a max length of 60, description has
a max length of 200 and each question has a max length of 150. Both of
these changes were implemented in the edit meeting and create meeting
files. They now match up to the Meeting model.