Created Models
I have created the Question model and Response model. They are both
registered to the admin site and they are all interconnected nicely.
I decided to keep them all in the director app since they are both
related to meetings. Maybe ill change it in the future. For now
it's fine.