Authentication Views
The authentication views have been completed for the most part.
The foundation has been set and the method for authentication
in this app will be session bassed via the session ID.
Thankfully DJANGO makes this easy for me. The authentication
system has yet to be fully tested and ISN't ready for prod
but it can be nicely built on top off now.