Updates to Meeting model
I have removed all previous migrations and rerun them with new updates to
the meeting model. The main update is that meeting models
now have a pk being their id. The id is a uuid4 field.
Secondly, ive added an access code feature. So that meetings
will be joined via the user implementing the proper meeting id.
That is all from me.