Live host and participant is added
The starting rendering pages have been made and
connected to the application. The pages look dull
and barebone, but it is well setup for testing.
For now we will focus on implementing the live chat logic,
ensuring that it works well and efficiently. Then,
we will make it look nice at the very very end of production.
Let's begin.