Meeting Host
Well, the websocket is now connecting to the channels.
Before there was an issue where authentication kept failing.
I found a temp fix to work around it in development. However in production
it will have to be implemented properly. But for now
we can keep working on the websocket now.