DOCX creation is complete
The standard docx file creation has been made,
additionally the way the code is structured is so that now
all types of files can be made. The data validation is stored in utils.py,
then utils.py is used in a generator file to create that file and send it
to the user.